refactor: Error타입 및 Error 핸들링  추가, 파라미터 명사형으로 변경 추가
