+++ JuiceMaker/JuiceMaker/Model/FruitResultType.swift
@@ -1,0 +1,13 @@
+//
+//  FruitResultType.swift
+//  JuiceMaker
+//
+//  Created by 김민제 on 1/24/24.
+//
+
+import Foundation
+
+enum FruitResultType: Error {
+    case success
+    case outOfStock
+}
