+++ JuiceMaker/JuiceMaker/Model/ResultType.swift
@@ -1,0 +1,13 @@
+//
+//  ResultType.swift
+//  JuiceMaker
+//
+//  Created by 김민제 on 1/24/24.
+//
+
+import Foundation
+
+enum ResultType: Error {
+    case success(String)
+    case failure(String)
+}
